requirements.txt
redis==5.0.1
celery==5.3.4
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai>=1.86.0
python-dotenv==1.0.0
sse-starlette==2.1.0
jinja2==3.1.4
pydantic-settings==2.0.3
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
requests